backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
supabase==2.4.6
pydantic==2.7.1
pydantic-settings==2.2.1

# Resume parsing
pdfminer.six==20221105
python-docx==1.1.2
spacy==3.7.4

# Matching engine
scikit-learn==1.4.2
numpy==1.26.4

# Scraping
playwright==1.44.0

# Email
resend==2.0.0

# Scheduling / utilities
python-jose[cryptography]==3.3.0
